Added Browser-router and routes
