+++ packages/frontend/src/pages/DetailPage.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+
+export default function Home() {
+    return (
+        <div>HELP</div>
+    )
+}
